requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==2023.10
cryptography==41.0.7
defusedxml==0.7.1
Django==4.2.7
django-allauth==0.59.0
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-filter==23.5
django-formtools==2.5.1
djangorestframework==3.14.0
idna==3.6
itypes==1.2.0
Jinja2==3.1.3
Markdown==3.5.2
MarkupSafe==2.1.5
numpy==2.2.4
oauthlib==3.2.2
pandas==2.2.3
pandas-stubs==2.2.3.250527
Pillow==10.1.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
qrcode==7.4.2
reportlab==4.0.7
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
types-pytz==2023.3.1.1
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
django-simple-history==3.4.0
WeasyPrint==60.1
cssselect2==0.7.0
tinycss2==1.2.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
mysqlclient==2.2.0
python-decouple==3.8
redis==5.0.1
django-redis==5.4.0
openpyxl==3.1.2

